# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local-first Markdown Kanban Board backend API.",
        "title": "Jotter API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Simon",
            "url": "https://github.com/simon123h/jotter"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/simon123h/jotter/blob/main/LICENSE.md"
        },
        "version": "2.9.1"
    },
    "basePath": "/api",
    "paths": {
        "/projects": {
            "get": {
                "description": "Get a list of all active projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/project.Response"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new workspace project board with default buckets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Create a new project",
                "parameters": [
                    {
                        "description": "Create Project Request",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project.Create"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/project.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{project_id}": {
            "put": {
                "description": "Update project metadata (title, clean period, or Git remote)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Update a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Project Request",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project.Update"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/project.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a project board and remove its local database cache index entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "description": "Get global user settings for the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get user settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/settings.AppSettings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Update and save global user settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Save user settings",
                "parameters": [
                    {
                        "description": "User Settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/settings.AppSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/system/sync": {
            "post": {
                "description": "Trigger a complete bidirectional sync. It parses task markdown files to rebuild the SQLite DB index, and fetches/merges/pushes active Git remotes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Synchronize system index and Git",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
        "project.Create": {
            "type": "object",
            "properties": {
                "done_clean_period": {
                    "type": "integer"
                },
                "git_remote": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "project.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "done_clean_period": {
                    "type": "integer"
                },
                "git_remote": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "project.Update": {
            "type": "object",
            "properties": {
                "done_clean_period": {
                    "type": "integer"
                },
                "git_remote": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "settings.AppSettings": {
            "type": "object",
            "properties": {
                "currentTheme": {
                    "type": "string"
                },
                "gitRemoteUrl": {
                    "type": "string"
                },
                "hideAddTaskButton": {
                    "type": "boolean"
                },
                "hideArchiveColumn": {
                    "type": "boolean"
                },
                "hideDoneColumn": {
                    "type": "boolean"
                },
                "isSidebarOpen": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "pinnedProjectIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "projectMru": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "sortBy": {
                    "type": "string"
                },
                "thresholdDays": {
                    "type": "integer"
                },
                "windowHeight": {
                    "type": "integer"
                },
                "windowMaximized": {
                    "type": "boolean"
                },
                "windowWidth": {
                    "type": "integer"
                },
                "windowX": {
                    "type": "integer"
                },
                "windowY": {
                    "type": "integer"
                }
            }
        },
...
    }
}